mydjangoapp/requirements.txt
Django==1.8.2
celery==3.1.18
djangorestframework==3.0.3
psycopg2==2.5.4
redis==2.10.3
python3-pika==0.9.14